requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.5.3
Pillow==9.4.0
plotly==5.9.0
scikit_learn==1.2.2
statsmodels==0.13.2
streamlit==1.16.0
